EmulatorService was unnecessary abstraction
