+++ src/scripts/emulator-view.ts
@@ -1,10 +1,12 @@
 import $ from 'jquery'
 import { bestCanvas } from 'jsbeeb/canvas'
-import { EmulatorCanvas } from './emulator'
+import { Emulator, EmulatorCanvas } from './emulator'
 import { Model } from 'jsbeeb/models'
 import { CustomAudioHandler } from './custom-audio-handler'
-import { BehaviorSubject } from 'rxjs'
-import { EmulatorService } from './emulator-service'
+import { BehaviorSubject, Observable, distinctUntilChanged } from 'rxjs'
+import { DisplayMode, getDisplayModeInfo } from './display-modes'
+import { notifyHost } from './vscode'
+import { ClientCommand } from '../types/shared/messages'
 
 const audioFilterFreq = 7000
 const audioFilterQ = 5
@@ -17,7 +19,11 @@ export class EmulatorView {
   canvas: EmulatorCanvas
   model: Model | undefined
   audioHandler: CustomAudioHandler
-  emulatorService: EmulatorService
+  emulator: Emulator | undefined // Dont hold references to the emulator, it may be paused and destroyed
+
+  // observables
+  private _displayMode$ = new BehaviorSubject<DisplayMode>(null)
+  readonly displayMode$: Observable<DisplayMode>
 
   fullscreen = new BehaviorSubject(false)
 
@@ -34,13 +40,11 @@ export class EmulatorView {
     this.canvas = bestCanvas(screen[0])
 
     // forward key events to emulator
-    screen.on('keyup', (event: JQuery.Event) =>
-      this.emulatorService.emulator?.onKeyUp(event),
-    )
+    screen.on('keyup', (event: JQuery.Event) => this.emulator?.onKeyUp(event))
     screen.on('keydown', (event: JQuery.Event) =>
-      this.emulatorService.emulator?.onKeyDown(event),
+      this.emulator?.onKeyDown(event),
     )
-    screen.on('blur', () => this.emulatorService.emulator?.clearKeys())
+    screen.on('blur', () => this.emulator?.clearKeys())
 
     // create webview audio driver
     this.audioHandler = new CustomAudioHandler(
@@ -54,17 +58,61 @@ export class EmulatorView {
     // little to get a reliable indication.
     window.setTimeout(() => this.audioHandler.checkStatus(), 1000)
 
-    this.emulatorService = new EmulatorService(this.canvas, this.audioHandler)
+    // setup observables
+    this.displayMode$ = this._displayMode$.pipe(distinctUntilChanged())
   }
 
   async initialise() {
     await this.audioHandler.initialise()
   }
 
-  async boot(model: Model) {
-    await this.emulatorService.boot(model)
-    this.showTestCard(this.emulatorService.emulator === undefined)
+  get displayMode(): DisplayMode | null {
+    return this._displayMode$.value
   }
+
+  async boot(model: Model) {
+    this.model = model
+    try {
+      if (this.emulator) {
+        this.emulator.shutdown()
+        this._displayMode$.next(null)
+      }
+
+      this.emulator = new Emulator(model, this.canvas, this.audioHandler)
+      await this.emulator.initialise()
+      notifyHost({ command: ClientCommand.EmulatorReady })
+
+      const discUrl = window.JSBEEB_DISC
+      await this.emulator.loadDisc(discUrl)
+      // if (discUrl) {
+      // 	const fdc = this.emulator.cpu.fdc;
+      // 	const discData = await utils.defaultLoadData(discUrl);
+      // 	const discImage = new BaseDisc(fdc, 'disc', discData, () => {});
+      // 	this.emulator.cpu.fdc.loadDisc(0, discImage);
+      // }
+      this.emulator.start()
+      // will automatically unsubscribe when emulator is shutdown
+      this.emulator.emulatorUpdate$.subscribe((emulator) =>
+        this.onEmulatorUpdate(emulator),
+      )
+    } catch (e) {
+      // this.showTestCard(true)
+      notifyHost({ command: ClientCommand.Error, text: (e as Error).message })
+      this.emulator = undefined
+    }
+    this.showTestCard(this.emulator === undefined)
+  }
+
+  private onEmulatorUpdate(_emulator: Emulator) {
+    // update display mode
+    const mode = this.emulator?.cpu.readmem(0x0355) ?? 255
+    this._displayMode$.next(getDisplayModeInfo(mode))
+  }
+
+  // async boot(model: Model) {
+  //   await this.emulatorService.boot(model)
+  //   this.showTestCard(this.emulatorService.emulator === undefined)
+  // }
 
   // async boot(model: Model) {
   //   this.model = model
@@ -99,7 +147,7 @@ export class EmulatorView {
   toggleFullscreen() {
     const isFullScreen = !this.fullscreen.value
     this.fullscreen.next(isFullScreen)
-    this.emulatorService.emulator?.setFullScreen(isFullScreen)
+    this.emulator?.setFullScreen(isFullScreen)
   }
 
   showTestCard(show: boolean = true) {
